[#233] fix UserPermmissions/UserPermission/Permissions confusion
