+++ app/models/user.rb
@@ -1,6 +1,5 @@
 class User < ActiveRecord::Base
   include ChatUser
-  include UserPermissions
 
   validates :sf_guard_user_id, presence: true, uniqueness: true
   validates :email, presence: true, uniqueness: { case_sensitive: false }
@@ -116,5 +115,47 @@ class User < ActiveRecord::Base
   def legacy_check_password(password)
     Digest::SHA1.hexdigest(sf_guard_user.salt + password) == sf_guard_user.password
   end
+
+  ###############
+  # Permissions #
+  ###############
+
+  def legacy_permissions
+    sf_guard_user.permissions
+  end
+
+  def has_legacy_permission(name)
+    sf_guard_user.has_permission(name)
+  end
+
+  def admin?
+    has_legacy_permission 'admin'
+  end
+
+  def importer?
+    has_legacy_permission 'importer'
+  end
+
+  def bulker?
+    has_legacy_permission 'bulker'
+  end
+
+  def merger?
+    has_legacy_permission 'merger'
+  end
+
+  def create_default_permissions
+    unless has_legacy_permission('contributor')
+      SfGuardUserPermission.create(permission_id: 2, user_id: sf_guard_user.id)
+    end
+    unless has_legacy_permission('editor')
+      SfGuardUserPermission.create(permission_id: 3, user_id: sf_guard_user.id)
+    end
+  end
+
+  def permissions
+    @permissions ||= Permissions.new(self)
+  end
   
+
 end
